change budget to float

+++ src/main/java/com/example/travelrecs/model/Trip.java
@@ -20,6 +20,6 @@ public class Trip {
 
     private Date startDate;
     private Date endDate;
-    private BigDecimal totalBudget;
-    private BigDecimal curBudget;
+    private Float totalBudget;
+    private Float curBudget;
 }
